+++ src/extension.js
@@ -1,9 +1,11 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 var vscode = require("vscode");
+const _ = require("lodash");
 
 // topper specific import
-const Topper = require("./topper.js");
+const topper = require("./topper");
+const topper2 = require("./topperv2");
 
 /**
  * Invoked when the extension is activated.
@@ -11,12 +13,26 @@ const Topper = require("./topper.js");
  */
 function activate(context) {
   console.log('Congratulations, your extension "topper" is now active!');
+  //==============================================================================================
   var addTopHeader = vscode.commands.registerCommand("topper.addTopHeader", function() {
     console.log("Topper is going to add header to the active file!");
-    Topper.topper.addTopHeader();
+    topper.addTopHeader();
     console.log("Your header has been successfully added to the file!");
   });
   context.subscriptions.push(addTopHeader);
+  //==============================================================================================
+  let profiles = vscode.workspace
+    .getConfiguration("topper")
+    .get("customTemplateParameters");
+  _.forEach(profiles, p =>
+    context.subscriptions.push(
+      vscode.commands.registerCommand(
+        `topper.addTopHeader.${Object.getOwnPropertyNames(p)[0]}`,
+        () => topper2.addTopHeaderForProfile(Object.getOwnPropertyNames(p)[0])
+      )
+    )
+  );
+  //==============================================================================================
 }
 
 /**